sai database, sua lai model, ve lai database

+++ ProductManager/Models/Admin.cs
@@ -7,7 +7,7 @@ namespace ProductManager.Models
 {
     public partial class Admin
     {
-        public short AdminId { get; set; }
+        public int AdminId { get; set; }
         public string UserName { get; set; }
         public string Password { get; set; }
         public string FullName { get; set; }
